LPS-59489 Apply changes
