+++ modules/apps/announcements/announcements-web/src/main/java/com/liferay/announcements/web/messaging/CheckEntryMessageListener.java
@@ -15,6 +15,8 @@
 package com.liferay.announcements.web.messaging;
 
 import com.liferay.portal.kernel.messaging.BaseSchedulerEntryMessageListener;
+import com.liferay.portal.kernel.messaging.Destination;
+import com.liferay.portal.kernel.messaging.DestinationNames;
 import com.liferay.portal.kernel.messaging.Message;
 import com.liferay.portal.kernel.module.framework.ModuleServiceLifecycle;
 import com.liferay.portal.kernel.scheduler.SchedulerEngineHelper;
@@ -45,7 +47,8 @@ public class CheckEntryMessageListener
 				PropsValues.ANNOUNCEMENTS_ENTRY_CHECK_INTERVAL,
 				TimeUnit.MINUTE));
 
-		_schedulerEngineHelper.register(this, schedulerEntryImpl);
+		_schedulerEngineHelper.register(
+			this, schedulerEntryImpl, DestinationNames.SCHEDULER_DISPATCH);
 	}
 
 	@Deactivate
@@ -63,6 +66,13 @@ public class CheckEntryMessageListener
 		AnnouncementsEntryLocalService announcementsEntryLocalService) {
 
 		_announcementsEntryLocalService = announcementsEntryLocalService;
+	}
+
+	@Reference(
+		target = "(destination.name=" + DestinationNames.SCHEDULER_DISPATCH + ")",
+		unbind = "-"
+	)
+	protected void setDestination(Destination destination) {
 	}
 
 	@Reference(target = ModuleServiceLifecycle.PORTAL_INITIALIZED, unbind = "-")
