+++ modules/apps/blogs/blogs-web/src/main/java/com/liferay/blogs/web/messaging/LinkbackMessageListener.java
@@ -18,6 +18,8 @@ import aQute.bnd.annotation.metatype.Configurable;
 
 import com.liferay.blogs.configuration.BlogsConfiguration;
 import com.liferay.portal.kernel.messaging.BaseSchedulerEntryMessageListener;
+import com.liferay.portal.kernel.messaging.Destination;
+import com.liferay.portal.kernel.messaging.DestinationNames;
 import com.liferay.portal.kernel.messaging.Message;
 import com.liferay.portal.kernel.module.framework.ModuleServiceLifecycle;
 import com.liferay.portal.kernel.scheduler.SchedulerEngineHelper;
@@ -54,7 +56,8 @@ public class LinkbackMessageListener extends BaseSchedulerEntryMessageListener {
 				getEventListenerClass(), getEventListenerClass(),
 				_blogsConfiguration.linkbackJobInterval(), TimeUnit.MINUTE));
 
-		_schedulerEngineHelper.register(this, schedulerEntryImpl);
+		_schedulerEngineHelper.register(
+			this, schedulerEntryImpl, DestinationNames.SCHEDULER_DISPATCH);
 	}
 
 	@Deactivate
@@ -67,6 +70,13 @@ public class LinkbackMessageListener extends BaseSchedulerEntryMessageListener {
 		_linkbackConsumer.verifyNewTrackbacks();
 
 		LinkbackProducerUtil.sendQueuedPingbacks();
+	}
+
+	@Reference(
+		target = "(destination.name=" + DestinationNames.SCHEDULER_DISPATCH + ")",
+		unbind = "-"
+	)
+	protected void setDestination(Destination destination) {
 	}
 
 	@Reference(target = ModuleServiceLifecycle.PORTAL_INITIALIZED, unbind = "-")
