+++ modules/apps/export-import/export-import-web/src/main/java/com/liferay/exportimport/web/messaging/DraftExportImportConfigurationMessageListener.java
@@ -25,6 +25,8 @@ import com.liferay.portal.kernel.dao.orm.PropertyFactoryUtil;
 import com.liferay.portal.kernel.dao.orm.QueryUtil;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.messaging.BaseSchedulerEntryMessageListener;
+import com.liferay.portal.kernel.messaging.Destination;
+import com.liferay.portal.kernel.messaging.DestinationNames;
 import com.liferay.portal.kernel.messaging.Message;
 import com.liferay.portal.kernel.module.framework.ModuleServiceLifecycle;
 import com.liferay.portal.kernel.scheduler.SchedulerEngineHelper;
@@ -65,7 +67,8 @@ public class DraftExportImportConfigurationMessageListener
 					DRAFT_EXPORT_IMPORT_CONFIGURATION_CHECK_INTERVAL,
 				TimeUnit.HOUR));
 
-		_schedulerEngineHelper.register(this, schedulerEntryImpl);
+		_schedulerEngineHelper.register(
+			this, schedulerEntryImpl, DestinationNames.SCHEDULER_DISPATCH);
 	}
 
 	@Deactivate
@@ -156,6 +159,13 @@ public class DraftExportImportConfigurationMessageListener
 		BackgroundTaskLocalService backgroundTaskLocalService) {
 
 		_backgroundTaskLocalService = backgroundTaskLocalService;
+	}
+
+	@Reference(
+		target = "(destination.name=" + DestinationNames.SCHEDULER_DISPATCH + ")",
+		unbind = "-"
+	)
+	protected void setDestination(Destination destination) {
 	}
 
 	@Reference(unbind = "-")
