+++ modules/apps/journal/journal-web/src/main/java/com/liferay/journal/web/messaging/CheckArticleMessageListener.java
@@ -18,6 +18,8 @@ import com.liferay.journal.service.JournalArticleLocalService;
 import com.liferay.journal.upgrade.JournalServiceUpgrade;
 import com.liferay.journal.web.configuration.JournalWebConfigurationValues;
 import com.liferay.portal.kernel.messaging.BaseSchedulerEntryMessageListener;
+import com.liferay.portal.kernel.messaging.Destination;
+import com.liferay.portal.kernel.messaging.DestinationNames;
 import com.liferay.portal.kernel.messaging.Message;
 import com.liferay.portal.kernel.module.framework.ModuleServiceLifecycle;
 import com.liferay.portal.kernel.scheduler.SchedulerEngineHelper;
@@ -46,7 +48,8 @@ public class CheckArticleMessageListener
 				getEventListenerClass(), getEventListenerClass(),
 				JournalWebConfigurationValues.CHECK_INTERVAL, TimeUnit.MINUTE));
 
-		_schedulerEngineHelper.register(this, schedulerEntryImpl);
+		_schedulerEngineHelper.register(
+			this, schedulerEntryImpl, DestinationNames.SCHEDULER_DISPATCH);
 	}
 
 	@Deactivate
@@ -57,6 +60,13 @@ public class CheckArticleMessageListener
 	@Override
 	protected void doReceive(Message message) throws Exception {
 		_journalArticleLocalService.checkArticles();
+	}
+
+	@Reference(
+		target = "(destination.name=" + DestinationNames.SCHEDULER_DISPATCH + ")",
+		unbind = "-"
+	)
+	protected void setDestination(Destination destination) {
 	}
 
 	@Reference(unbind = "-")
