+++ modules/apps/message-boards/message-boards-web/src/main/java/com/liferay/message/boards/web/messaging/ExpireBanMessageListener.java
@@ -18,6 +18,8 @@ import aQute.bnd.annotation.metatype.Configurable;
 
 import com.liferay.message.boards.configuration.MBConfiguration;
 import com.liferay.portal.kernel.messaging.BaseSchedulerEntryMessageListener;
+import com.liferay.portal.kernel.messaging.Destination;
+import com.liferay.portal.kernel.messaging.DestinationNames;
 import com.liferay.portal.kernel.messaging.Message;
 import com.liferay.portal.kernel.module.framework.ModuleServiceLifecycle;
 import com.liferay.portal.kernel.scheduler.SchedulerEngineHelper;
@@ -53,7 +55,8 @@ public class ExpireBanMessageListener
 				getEventListenerClass(), getEventListenerClass(),
 				_mbConfiguration.expireBanJobInterval(), TimeUnit.MINUTE));
 
-		_schedulerEngineHelper.register(this, schedulerEntryImpl);
+		_schedulerEngineHelper.register(
+			this, schedulerEntryImpl, DestinationNames.SCHEDULER_DISPATCH);
 	}
 
 	@Deactivate
@@ -64,6 +67,13 @@ public class ExpireBanMessageListener
 	@Override
 	protected void doReceive(Message message) throws Exception {
 		_mbBanLocalService.expireBans();
+	}
+
+	@Reference(
+		target = "(destination.name=" + DestinationNames.SCHEDULER_DISPATCH + ")",
+		unbind = "-"
+	)
+	protected void setDestination(Destination destination) {
 	}
 
 	@Reference(unbind = "-")
