+++ modules/apps/trash/trash-web/src/main/java/com/liferay/trash/web/messaging/CheckEntryMessageListener.java
@@ -15,6 +15,8 @@
 package com.liferay.trash.web.messaging;
 
 import com.liferay.portal.kernel.messaging.BaseSchedulerEntryMessageListener;
+import com.liferay.portal.kernel.messaging.Destination;
+import com.liferay.portal.kernel.messaging.DestinationNames;
 import com.liferay.portal.kernel.messaging.Message;
 import com.liferay.portal.kernel.module.framework.ModuleServiceLifecycle;
 import com.liferay.portal.kernel.scheduler.SchedulerEngineHelper;
@@ -43,7 +45,8 @@ public class CheckEntryMessageListener
 				getEventListenerClass(), getEventListenerClass(),
 				PropsValues.TRASH_ENTRY_CHECK_INTERVAL, TimeUnit.MINUTE));
 
-		_schedulerEngineHelper.register(this, schedulerEntryImpl);
+		_schedulerEngineHelper.register(
+			this, schedulerEntryImpl, DestinationNames.SCHEDULER_DISPATCH);
 	}
 
 	@Deactivate
@@ -54,6 +57,13 @@ public class CheckEntryMessageListener
 	@Override
 	protected void doReceive(Message message) throws Exception {
 		_trashEntryLocalService.checkEntries();
+	}
+
+	@Reference(
+		target = "(destination.name=" + DestinationNames.SCHEDULER_DISPATCH + ")",
+		unbind = "-"
+	)
+	protected void setDestination(Destination destination) {
 	}
 
 	@Reference(target = ModuleServiceLifecycle.PORTAL_INITIALIZED, unbind = "-")
