+++ modules/portal/portal-ldap/src/main/java/com/liferay/portal/ldap/internal/messaging/UserImportMessageListener.java
@@ -14,10 +14,13 @@
 
 package com.liferay.portal.ldap.internal.messaging;
 
-import com.liferay.portal.kernel.messaging.BaseMessageListener;
+import com.liferay.portal.kernel.messaging.BaseSchedulerEntryMessageListener;
 import com.liferay.portal.kernel.messaging.Destination;
 import com.liferay.portal.kernel.messaging.Message;
-import com.liferay.portal.kernel.messaging.MessageListener;
+import com.liferay.portal.kernel.scheduler.SchedulerEngineHelper;
+import com.liferay.portal.kernel.scheduler.TimeUnit;
+import com.liferay.portal.kernel.scheduler.TriggerFactory;
+import com.liferay.portal.kernel.scheduler.TriggerFactoryUtil;
 import com.liferay.portal.ldap.configuration.ConfigurationProvider;
 import com.liferay.portal.ldap.exportimport.configuration.LDAPImportConfiguration;
 import com.liferay.portal.model.Company;
@@ -26,20 +29,41 @@ import com.liferay.portal.service.CompanyLocalService;
 
 import java.util.List;
 
+import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Deactivate;
+import org.osgi.service.component.annotations.Modified;
 import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Shuyang Zhou
  */
-@Component(
-	immediate = true,
-	property = {
-		"destination.name=" + DestinationNames.SCHEDULED_USER_LDAP_IMPORT
-	},
-	service = MessageListener.class
-)
-public class UserImportMessageListener extends BaseMessageListener {
+@Component(immediate = true, service = UserImportMessageListener.class)
+public class UserImportMessageListener
+	extends BaseSchedulerEntryMessageListener {
+
+	@Activate
+	@Modified
+	protected void activate() {
+		LDAPImportConfiguration ldapImportConfiguration =
+			_ldapImportConfigurationProvider.getConfiguration(0L);
+
+		int interval = ldapImportConfiguration.importInterval();
+
+		schedulerEntryImpl.setTrigger(
+			TriggerFactoryUtil.createTrigger(
+				getEventListenerClass(), getEventListenerClass(), interval,
+				TimeUnit.MINUTE));
+
+		_schedulerEngineHelper.register(
+			this, schedulerEntryImpl,
+			DestinationNames.SCHEDULED_USER_LDAP_IMPORT);
+	}
+
+	@Deactivate
+	protected void deactivate() {
+		_schedulerEngineHelper.unregister(this);
+	}
 
 	@Override
 	protected void doReceive(Message message) throws Exception {
@@ -92,9 +116,21 @@ public class UserImportMessageListener extends BaseMessageListener {
 		_ldapUserImporter = ldapUserImporter;
 	}
 
+	@Reference(unbind = "-")
+	protected void setSchedulerEngineHelper(
+		SchedulerEngineHelper schedulerEngineHelper) {
+
+		_schedulerEngineHelper = schedulerEngineHelper;
+	}
+
+	@Reference(unbind = "-")
+	protected void setTriggerFactory(TriggerFactory triggerFactory) {
+	}
+
 	private CompanyLocalService _companyLocalService;
 	private ConfigurationProvider<LDAPImportConfiguration>
 		_ldapImportConfigurationProvider;
 	private LDAPUserImporter _ldapUserImporter;
+	private SchedulerEngineHelper _schedulerEngineHelper;
 
 }
